requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.21
Werkzeug==0.9.4
gunicorn==18.0
itsdangerous==0.24
wsgiref==0.1.2
